client/draw - pruning events, adding CONSUME_HANDLES
dealing with events is becoming more and more burdensome. I like the idea of having a central list of events, but if you need to add/remove listeners often it becomes a HUGE freaking hassle.  I really want to be able to async a lot of this shit if I have to, but it’s getting weirdly complicated to do so.